Update storage-files-active-directory-overview.md
Misleading or incomplete statement.
"Your domain-joined VM must reside in the same virtual network (VNET) as Azure AD Domain Services."

The statement gives the impression that the VM's have to share the same network space as the managed domain, There are instances in which AADDS is not subneted and it resides on its own VNet and connects to other networks to act as authority. https://docs.microsoft.com/en-us/azure/active-directory-domain-services/network-considerations states that the managed domain has to be deployed and isolated to it's own subnet this may be misinterpreted and lead to believe that this means same subnet, also AADDS can offer its capabillities and act as authority to other networks via, peering, exprees route or vpn (extending its range across multiple networks. It is not clearly estated whether or not these scenarios are supported by the feature.